Reorganized API views
Moved the views into resource-specific modules.

ECOM-6501
